journal_server/requirements/local.txt
-r base.txt


Werkzeug[watchdog]==2.3.7 
ipdb==0.13.13  
watchfiles==0.20.0  


# Testing
# ------------------------------------------------------------------------------
mypy==1.10.0 
django-stubs[compatible-mypy]==5.0.2 
pytest==7.4.2 
pytest-sugar==0.9.7 
djangorestframework-stubs[compatible-mypy]==3.15.0


# Code quality
# ------------------------------------------------------------------------------
flake8==6.1.0 
flake8-isort==6.1.0 
coverage==7.3.1  
black==23.9.1  
djlint==1.32.1  
pylint-django==2.5.3 
pylint-celery==0.3  
pre-commit==3.4.0  

# Django
# ------------------------------------------------------------------------------
factory-boy==3.3.0  

# django-debug-toolbbasear==4.4.2  
django-extensions==3.2.3  
django-coverage-plugin==3.1.0  
pytest-django==4.5.2  